Move common implementation to `internal` package
