+++ src/commonImplementation/kotlin/StringBuilderExtensions.kt
@@ -1,5 +1,6 @@
 package de.cketti.codepoints
 
+import de.cketti.codepoints.internal.CommonStringBuilderFunctions
 import kotlin.text.StringBuilder
 
 actual fun StringBuilder.appendCodePoint(codePoint: Int): StringBuilder = apply {
